Change vizzini frequency to every 10 mins
Load seems too high on Thanos cell, so giving it more breathing space

Vizzini takes ~3 min to run, so running every 10 should give the cell
more than enough breathing space. We've since found that the cell's CPU
is completely slammed during the Vizzini run, so we think that things
are going wrong 'in the moment', rather than the issue being related to
the Vizzini run frequency.

[#145276295]

Signed-off-by: Gareth Clay <gclay@pivotal.io>